requirements.txt
fastapi[all]

# Tools for handling JWTs
python-jose[cryptography]
passlib[bcrypt]

# Database ORM and migrations
sqlalchemy
alembic

# Support for postgres
psycopg2
